+++ packages/charting/src/common/styles.js
@@ -1,16 +1,15 @@
-import green from '@material-ui/core/colors/green';
-import orange from '@material-ui/core/colors/orange';
+import { color } from '@pie-lib/render-ui';
 
 export const disabled = (key = 'fill') => ({
-  [key]: `var(--graph-disabled, ${'rgba(0,0,0,0.2)'})`,
+  [key]: `var(--graph-disabled, ${color.disabled()})`,
   pointerEvents: 'none'
 });
 
 export const correct = (key = 'fill') => ({
-  [key]: green[500],
+  [key]: color.correct(),
   pointerEvents: 'none'
 });
 export const incorrect = (key = 'fill') => ({
-  [key]: orange[500],
+  [key]: color.incorrect(),
   pointerEvents: 'none'
 });