+++ packages/mask-markup/src/components/dropdown.jsx
@@ -3,6 +3,8 @@ import PropTypes from 'prop-types';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import CorrectInput from './correct-input';
+import { withStyles } from '@material-ui/core/styles';
+import { color } from '@pie-lib/render-ui';
 
 class Dropdown extends React.Component {
   static propTypes = {
@@ -10,6 +12,7 @@ class Dropdown extends React.Component {
     value: PropTypes.string,
     disabled: PropTypes.bool,
     onChange: PropTypes.func,
+    classes: PropTypes.object,
     correct: PropTypes.bool,
     choices: PropTypes.arrayOf(
       PropTypes.shape({ value: PropTypes.string, label: PropTypes.string })
@@ -18,10 +21,20 @@ class Dropdown extends React.Component {
   };
 
   render() {
-    const { id, correct, disabled, value, onChange, choices, showCorrectAnswer } = this.props;
+    const {
+      classes,
+      id,
+      correct,
+      disabled,
+      value,
+      onChange,
+      choices,
+      showCorrectAnswer
+    } = this.props;
 
     return (
       <Select
+        classes={{ root: classes.root, icon: classes.icon, selectMenu: classes.selectMenu }}
         disabled={disabled}
         value={value || ''}
         input={<CorrectInput correct={showCorrectAnswer || correct} />}
@@ -34,7 +47,11 @@ class Dropdown extends React.Component {
         }}
       >
         {(choices || []).map((c, index) => (
-          <MenuItem key={`${c.label}-${index}`} value={c.value}>
+          <MenuItem
+            classes={{ root: classes.menuRoot, selected: classes.selected }}
+            key={`${c.label}-${index}`}
+            value={c.value}
+          >
             <span
               ref={ref => {
                 if (ref) {
@@ -49,4 +66,38 @@ class Dropdown extends React.Component {
   }
 }
 
-export default Dropdown;
+const styles = () => ({
+  root: {
+    color: color.text(),
+    backgroundColor: color.background(),
+    border: '1px solid gray',
+    '& ul': {
+      color: color.text(),
+      backgroundColor: color.background()
+    }
+  },
+  selectMenu: {
+    backgroundColor: color.background()
+  },
+  icon: {
+    color: color.text()
+  },
+  selected: {
+    color: `${color.primary()} !important`,
+    backgroundColor: `${color.secondary()} !important`
+  },
+  menuRoot: {
+    color: color.primaryDark(),
+    backgroundColor: color.secondaryDark(),
+    '&:focus': {
+      color: color.primary(),
+      backgroundColor: color.secondary()
+    },
+    '&:hover': {
+      color: color.primary(),
+      backgroundColor: color.secondary()
+    }
+  }
+});
+
+export default withStyles(styles)(Dropdown);