+++ packages/charting/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.3.0](https://github.com/pie-framework/pie-lib/compare/@pie-lib/charting@4.2.22...@pie-lib/charting@4.3.0) (2020-10-07)
+
+
+### Features
+
+* **charting:** add color theme support ([606d45e](https://github.com/pie-framework/pie-lib/commit/606d45e))
+
+
+
+
+
 ## [4.2.22](https://github.com/pie-framework/pie-lib/compare/@pie-lib/charting@4.2.21...@pie-lib/charting@4.2.22) (2020-05-15)
 
 **Note:** Version bump only for package @pie-lib/charting