+++ packages/editable-html/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.13.0](https://github.com/pie-framework/pie-lib/compare/@pie-lib/editable-html@7.12.12...@pie-lib/editable-html@7.13.0) (2020-10-07)
+
+
+### Features
+
+* **editable-html:** color theme support added ([d7e32fd](https://github.com/pie-framework/pie-lib/commit/d7e32fd))
+* **editable-html:** initial commit for color treatment ([970c4d1](https://github.com/pie-framework/pie-lib/commit/970c4d1))
+
+
+
+
+
 ## [7.12.12](https://github.com/pie-framework/pie-lib/compare/@pie-lib/editable-html@7.12.11...@pie-lib/editable-html@7.12.12) (2020-09-18)
 
 **Note:** Version bump only for package @pie-lib/editable-html